+++ Rebus.AzureServiceBus.Tests/Factories/StandardAzureServiceBusTransportFactory.cs
@@ -87,17 +87,26 @@ namespace Rebus.AzureServiceBus.Tests.Factories
 
         public void CleanUp()
         {
+            _queuesToDelete.Keys.ForEach(DeleteQueue);
+        }
+
+        public static void DeleteQueue(string queueName)
+        {
             var namespaceManager = NamespaceManager.CreateFromConnectionString(ConnectionString);
 
-            _queuesToDelete.Keys.ForEach(queueName =>
+            if (!namespaceManager.QueueExists(queueName)) return;
+
+            Console.Write("Deleting ASB queue {0}...", queueName);
+
+            try
             {
-                if (!namespaceManager.QueueExists(queueName)) return;
-
-                Console.WriteLine("Deleting ASB queue {0}", queueName);
-
                 namespaceManager.DeleteQueue(queueName);
-            });
-        }
+                Console.WriteLine("OK!");
+            }
+            catch (MessagingEntityNotFoundException)
+            {
+                Console.WriteLine("OK (was not there)");   
+            }        }
 
         public static void DeleteTopic(string topic)
         {